Move apply_strip() to unpacker_utils
This way it can be used by the tar implementation.
